Refactor task modules under src/tasks
