+++ src/tasks/service.rs
@@ -13,9 +13,9 @@ use chrono::Utc;
 
 use crate::commands::common::is_process_running;
 use crate::commands::tasks::{collect_active_tasks, collect_archived_tasks};
-use crate::status::derive_active_state;
-use crate::storage::{LOG_FILE_NAME, TaskPaths, TaskStore};
-use crate::task::{TaskMetadata, TaskState};
+use crate::tasks::{
+    LOG_FILE_NAME, TaskMetadata, TaskPaths, TaskState, TaskStore, derive_active_state,
+};
 use crate::worker::launcher::{WorkerLaunchRequest, spawn_worker};
 
 const SHUTDOWN_TIMEOUT_SECS: u64 = 10;
